Adds a `preprocess_args` option to `convert_to_tensors`
Adds a mechanism to preprocess arguments before they are converted to tensors
so that we don't need to introduce helpers and use `skip_num_stack_entries`.
