+++ tripy/tests/integration/test_reshape.py
@@ -49,42 +49,3 @@ class TestReshape:
         a = tp.ones((2, 3, 4))
         out = tp.reshape(a, (2, a.shape[1], a.shape[2] / 2, -1))
         assert np.array_equal(cp.from_dlpack(out).get(), np.ones((2, 3, 2, 2), dtype=np.float32))
-
-
-class TestFlatten:
-    @pytest.mark.parametrize(
-        "shape, start_dim, end_dim, expected_shape",
-        [
-            ((2, 3, 4), 0, -1, (24,)),  # Flatten all dimensions
-            ((2, 3, 4), 1, -1, (2, 12)),  # Flatten dimensions 1 through end
-            ((2, 3, 4), 1, 2, (2, 12)),  # Flatten dimensions 1 through 2
-            ((2, 3, 4), 0, 1, (6, 4)),  # Flatten dimensions 0 through 1
-            ((2, 3, 4, 5), 1, 3, (2, 60)),  # Flatten dimensions 1 through 3
-        ],
-    )
-    def test_flatten(self, shape, start_dim, end_dim, expected_shape):
-        cp_a = cp.arange(np.prod(shape)).reshape(shape).astype(np.float32)
-        a = tp.Tensor(cp_a)
-        b = tp.flatten(a, start_dim=start_dim, end_dim=end_dim)
-        assert b.shape == list(expected_shape)
-        assert np.array_equal(cp.from_dlpack(b).get(), cp_a.reshape(expected_shape).get())
-
-    def test_flatten_invalid_dims(self):
-        shape = (2, 3, 4)
-        with pytest.raises(tp.TripyException, match="Invalid dimensions"):
-            a = tp.ones(shape)
-            # Invalid because end_dim < start_dim
-            tp.flatten(a, start_dim=2, end_dim=1)
-
-    def test_flatten_single_dim(self):
-        shape = (2, 3, 4)
-        a = tp.ones(shape)
-        # Flattening a single dimension should not change the output
-        b = tp.flatten(a, start_dim=1, end_dim=1)
-        assert b.shape == [2, 3, 4]
-        assert np.array_equal(cp.from_dlpack(b).get(), np.ones(shape, dtype=np.float32))
-
-    def test_flatten_with_unknown_dims(self):
-        a = tp.ones((2, 3, 4, 5))
-        b = tp.flatten(a, start_dim=1, end_dim=-1)
-        assert np.array_equal(cp.from_dlpack(b).get(), np.ones((2, 60), dtype=np.float32))
